+++ Projects/Home/Sources/AudioTrack/OverlayView.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+import hCoreUI
+
+struct OverlayView: View {
+    @ObservedObject var audioPlayer: AudioPlayer
+
+    let staplesMaskColor: some hColor = hTintColor.lavenderOne
+
+    var body: some View {
+        GeometryReader { geometry in
+            ZStack(alignment: .leading) {
+                Rectangle()
+                    .fill(staplesMaskColor)
+                    .frame(width: geometry.size.width * audioPlayer.progress)
+            }
+        }
+    }
+}
