+++ Projects/Home/Sources/AudioTrack/TrackPlayer.swift
@@ -1,8 +1,4 @@
-import AVFoundation
-import Combine
-import Foundation
 import SwiftUI
-import Swifter
 import hCore
 import hCoreUI
 
@@ -36,99 +32,5 @@ struct TrackPlayer: View {
                 audioPlayer.togglePlaying()
             }
         }
-    }
-}
-
-struct OverlayView: View {
-    @ObservedObject var audioPlayer: AudioPlayer
-
-    let staplesMaskColor: some hColor = hTintColor.lavenderOne
-
-    var body: some View {
-        GeometryReader { geometry in
-            ZStack(alignment: .leading) {
-                Rectangle()
-                    .fill(staplesMaskColor)
-                    .frame(width: geometry.size.width * audioPlayer.progress)
-            }
-        }
-    }
-}
-
-class AudioPlayer: NSObject, ObservableObject, AVAudioPlayerDelegate {
-    internal init(
-        url: URL,
-        isPlaying: Bool = false
-    ) {
-        self.url = url
-        self.isPlaying = isPlaying
-    }
-
-    let objectWillChange = PassthroughSubject<AudioPlayer, Never>()
-
-    var audioPlayer: AVPlayer?
-
-    let playerTimer = Timer.publish(every: 1 / 30, on: .main, in: .common)
-        .autoconnect()
-
-    let url: URL
-
-    private(set) var isPlaying: Bool = false {
-        didSet {
-            objectWillChange.send(self)
-        }
-    }
-
-    private(set) var progress: Double = 0 {
-        didSet {
-            objectWillChange.send(self)
-        }
-    }
-
-    func togglePlaying() {
-        isPlaying ? stopPlaying() : startPlaying()
-    }
-
-    private func startPlaying() {
-        let session = AVAudioSession.sharedInstance()
-
-        do {
-            try session.setCategory(.playback)
-            try session.overrideOutputAudioPort(AVAudioSession.PortOverride.speaker)
-            try session.setActive(true)
-        } catch {
-            print("Playing over the device's speakers failed")
-        }
-
-        let playerItem = AVPlayerItem(url: url)
-        audioPlayer = AVPlayer(playerItem: playerItem)
-
-        audioPlayer?
-            .addPeriodicTimeObserver(
-                forInterval: CMTime(value: 1, timescale: 1),
-                queue: .main,
-                using: { [weak self] time in
-                    guard let self = self else { return }
-                    if let item = self.audioPlayer?.currentItem, item.status == .readyToPlay {
-                        let duration = CMTimeGetSeconds(item.duration)
-                        let timeInFloat = CMTimeGetSeconds(time)
-                        self.progress = timeInFloat / duration
-
-                    }
-                }
-            )
-
-        audioPlayer?.actionAtItemEnd = .pause
-        audioPlayer?.play()
-        isPlaying = true
-    }
-
-    private func stopPlaying() {
-        audioPlayer?.pause()
-        isPlaying = false
-    }
-
-    func audioPlayerDidFinishPlaying(_ player: AVAudioPlayer, successfully flag: Bool) {
-        isPlaying = false
     }
 }
