--- Projects/hCore/Sources/DateFormatter+ISO8601.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-extension DateFormatter {
-    public static func withIso8601Format(_ format: String) -> DateFormatter {
-        let formatter = DateFormatter()
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.dateFormat = format
-        return formatter
-    }
-}
