+++ Projects/hCore/Sources/String+toDate.swift
@@ -7,4 +7,11 @@ extension String {
         formatter.dateFormat = "yyyy-MM-dd"
         return formatter.date(from: self)
     }
+    
+    public func localDateToIso8601Date(format: String) -> Date? {
+        let formatter = DateFormatter()
+        formatter.calendar = Calendar(identifier: .iso8601)
+        formatter.dateFormat = format
+        return formatter.date(from: self)
+    }
 }
